更新角色权限

+++ FytSoa.Core/Model/Cms/CmsLevel.cs
@@ -5,9 +5,9 @@
     {
         public int Id {get;set;}
         public string Name { get; set; }
-        public int Alipay { get; set; }
-        public int WxPay { get; set; }
-        public int OtherPay { get; set; }
+        public decimal Alipay { get; set; }
+        public decimal WxPay { get; set; }
+        public decimal OtherPay { get; set; }
         public bool Status { get; set; }
     }
 }
